Config validator work in progress
